GD-357: Parameterized test cases are not run from Godot Editor GdUnit4 inspector
# Why
When running C# tests from Godot Editor the parameterized tests are not executed.
This was happen since we introduce the test filter with introduction the test adapter.

# What
Replaced the invalid filter in the executor and load by `LoadTestFilter` the saved runner config to determine the included tests.
Added a small hack to the `FilterByTestCaseName` to work with "simplyfied" parameterized testnames
Added missing method `set_test_parameter_index` on CsNode called from GdScript.