GD-357: Parameterized test cases are not run from Godot Editor GdUnit4 inspector (#59)
# Why
When running C# tests from Godot Editor, the parameterized tests are not
executed. This was happened since we introduce the test filter with
introduction of the test adapter.

# What
Replaced the invalid filter in the executor and load by `LoadTestFilter`
the saved runner config to determine the included tests. Added a small
hack to the `FilterByTestCaseName` to work with "simplified"
parameterized test names.
Added missing method `set_test_parameter_index` on `CsNode` called from
`GdScript`.

set API version to 4.2.1.1