GD-90: Cleanup old runner configuration before run new tests (#93)
# Why
On Visual Studio, there are leftovers when press stop debug on a test
session. This results into a `The active test run was aborted. Reason:
Test host process crashed` This crash can't be handled by the test
executor because the process is lost.

# What
- added to `.gitignore` a rule do not push runner config files
- do delete now old runner config files if exist before write the new
one. This is just a workaround to clean up the working directory