GD-333: Fixing broken Signal assertion task handling (#346)
# Why
I want to verify a signal is not emitted in a given time frame, but it
breaks my before successful running test. The signal assert extension is
broken, the `WithTimeout` is not canceling the running task.

# What
There was many issues how the signal assertion is creating/cancel and
continuation the signal task.
- Complete reworked on the signal assert task handling
- Using a thread save map to store the task id and cancelation token to
handle `WithTimeout` cancelation the running signal task
- Removed redundant task recreation to have a stable task id for
cancelation
- Fixing signal assert exceptions creation using the correct stacktrace