GD-348: Fixing task handling using `AwaitMethod` with `WithTimeout` (#349)
# Why
When waiting for a method and specify a timeout, the original method
task is not canceled when the timeout is reached. This will result the
original task is still running, and the cancellation toke is not freed.

# What
- Rework on `GodotMethodAwaitable` task creation
- Using a thread save map to store the task id and cancellation token to
handle `WithTimeout` cancellation of the running task
- Removed redundant task recreation to have a stable task id for
cancellation
- Fixing `ExecutionTimeoutException` creation using the correct stack
trace


Bump to release candidate 1