GD-356: Fixing error: `Parameter "p_node" is null` when using SceneTree as emitter for AwaitSignalOn (#357)
# Why
When a user do await on the scene tree itself, an error is shown. `
await AwaitSignalOn(ctrl.GetTree(), SceneTree.SignalName.ProcessFrame);`
The issue is we determine is calling process (physic) frame on given
emitter by using ` occurs at `sceneTree?.Root.GetPathTo(emitter as
Node);` This will result in a `"p_node" is null` error because we can't
cast the emitter to a Node

# What
- Rework `DoesNodeProcessing` to handle node is inside tree to avoid the
error.